feat(ui): update landing page components for a more coherent look
Links didn't fit the design language I was going for very well. Added a
simple border and hover effect that ties the landing page together
better.

Updated the "coming soon" section to align better with the rest of the
landing page.

Moved the background animation to the landing page instead of the
layout. I only want the effect to show on the landing page instead of
everywhere.